Switch internal API over to new router
Hello! 👋

I have been working on a new Router package in my spare time. The goal
is to have it replace itty-router. I tried upgrading itty-router to the
latest version which is a lot better, but I've been hitting some
difficulties with it, and figured that if we'd go through this pain, we
might was well rewrite the router ourselves from scratch ¯\_(ツ)_/¯

Nothing against against itty-router! It's been great for us to get us
started, and has taken us far. And it's really tiny. However, the reason
it's tiny is that it barely includes any logic. This means we have been
reinventing the same wheel multiple times: how to handle authentication,
how to validate inputs, how to handle errors, how to do CORS, etc. Some
of that support code exists 3 times in different forms. Things work
subtly differently depending on where in the API you are.

My router takes a different approach: it includes batteries, and tries
to optimize for best practices. Here is a list of its guiding
principles. I plan to open source this router once I consider it fully
production-ready. The current package name's working title is serv, you
can find the code here. It's currently still just a private NPM
workspace inside this repo, to make iterating on it easier.

This PR does two things:

1. It adds the new router's implementation, see the packages/serv
   directory
2. It uses this new router for the /internal/* API subroutes, see the
   cleanup in src/api/internal/index.ts.

> Note:
> This means for /api/v1, /v2, and /v6 and /v7 sockets, and for the
> top-level routing, it's still using itty-router.

Once all other subroutes have been replaced as well, we can clean up
a lot more supporting functions, and drop the itty-router dependency.
I think eventually in the end it will be a net-removal of code.

First, I'd like to collect feedback in this PR, and get this pushed to
production.

The DX is pretty nice.

Original PR: https://github.com/liveblocks/liveblocks-cloudflare/pull/823